+++ doc/changes.rst
@@ -39,6 +39,20 @@ Added :func:`apsw.ext.ShowResourceUsage` for getting resource and
 SQLite usage in a block, and also use it for the shell :ref:`timer
 <shell-cmd-timer>` command.
 
+3.46.0.0
+========
+
+Adjusted `levels
+<https://docs.python.org/3/library/logging.html#levels>`__ in
+:func:`apsw.ext.log_sqlite` to be lower for some SQLite messages like
+`SQLITE_SCHEMA` and `SQLITE_NOTICE_RECOVER_WAL` (:issue:`518`)
+
+Previous source releases were signed with `PGP
+<https://en.wikipedia.org/wiki/Pretty_Good_Privacy>`__.  Starting with
+this release `Sigstore's <https://www.sigstore.dev/>`__ `cosign tool
+<https://docs.sigstore.dev/signing/quickstart/>`__ is used
+(:ref:`instructions <verifydownload>`). (:issue:`512`)
+
 3.45.3.0
 ========